+++ app/Listeners/Forms/NotifyFormSubmission.php
@@ -4,7 +4,7 @@ namespace App\Listeners\Forms;
 
 use App\Events\Forms\FormSubmitted;
 use App\Models\Integration\FormIntegration;
-use App\Service\Forms\Integrations\AbstractIntegrationHandler;
+use App\Integrations\Handlers\AbstractIntegrationHandler;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Queue\InteractsWithQueue;
 
@@ -35,9 +35,9 @@ class NotifyFormSubmission implements ShouldQueue
     ): AbstractIntegrationHandler {
         $integration = FormIntegration::getIntegration($formIntegration->integration_id);
         if ($integration && isset($integration['file_name']) && class_exists(
-            'App\Service\Forms\Integrations\\' . $integration['file_name']
+            'App\Integrations\Handlers\\' . $integration['file_name']
         )) {
-            $className = 'App\Service\Forms\Integrations\\' . $integration['file_name'];
+            $className = 'App\Integrations\Handlers\\' . $integration['file_name'];
             return new $className($event, $formIntegration, $integration);
         }
         throw new \Exception('Unknown Integration!');
